Moved ActorAlertManagement to controller package
